Address PR review feedback: deprecation warnings, clearer error, docs
- Emit DeprecationWarning (Python) / console.warn (JS) when the
  deprecated auto_pause/autoPause flag is set on beta_create.
- Clarify validation error message: 'autoResume can only be true when
  the resolved onTimeout is pause' (was misleading when explicit
  onTimeout='kill' took precedence over autoPause=true).
- Remove dual-purpose local variable in JS getLifecycle for clarity;
  inline the autoResume===true check.
- Replace non-standard ':deprecated:' docstring placement with
  idiomatic '**Deprecated**:' inline marker on auto_pause param.
- Consolidate duplicated get_lifecycle helper tests into a single
  tests/test_sandbox_api.py module (helper lives in shared code).
- Document in changeset that auto_resume=true + on_timeout=kill is now
  a hard error (previously silently accepted) and that JS now always
  sends autoPause in the create-sandbox body.
