+++ packages/python-sdk/e2b/sandbox_async/main.py
@@ -3,6 +3,7 @@ import json
 import logging
 import shlex
 import uuid
+import warnings
 from typing import Dict, List, Optional, Union, overload
 
 import httpx
@@ -551,13 +552,14 @@ class AsyncSandbox(SandboxApi):
         cls,
         template: Optional[str] = None,
         timeout: Optional[int] = None,
-        auto_pause: bool = False,
+        auto_pause: Optional[bool] = None,
         metadata: Optional[Dict[str, str]] = None,
         envs: Optional[Dict[str, str]] = None,
         secure: bool = True,
         allow_internet_access: bool = True,
         mcp: Optional[McpServer] = None,
         volume_mounts: Optional[SandboxAsyncVolumeMount] = None,
+        lifecycle: Optional[SandboxLifecycle] = None,
         **opts: Unpack[ApiParams],
     ) -> Self:
         """
@@ -569,19 +571,27 @@ class AsyncSandbox(SandboxApi):
 
         :param template: Sandbox template name or ID
         :param timeout: Timeout for the sandbox in **seconds**, default to 300 seconds. The maximum time a sandbox can be kept alive is 24 hours (86_400 seconds) for Pro users and 1 hour (3_600 seconds) for Hobby users.
-        :param auto_pause: Automatically pause the sandbox after the timeout expires. Defaults to `False`.
-            :deprecated: Use `lifecycle={"on_timeout": "pause"}` instead.
+        :param auto_pause: **Deprecated**: use ``lifecycle={"on_timeout": "pause"}`` instead.
+            Automatically pause the sandbox after the timeout expires. Defaults to ``False``.
         :param metadata: Custom metadata for the sandbox
         :param envs: Custom environment variables for the sandbox
         :param secure: Envd is secured with access token and cannot be used without it, defaults to `True`.
         :param allow_internet_access: Allow sandbox to access the internet, defaults to `True`.
         :param mcp: MCP server to enable in the sandbox
         :param volume_mounts: Dictionary mapping mount paths to AsyncVolume instances or volume names
+        :param lifecycle: Sandbox lifecycle configuration; controls timeout behavior and auto-resume.
 
         :return: A Sandbox instance for the new sandbox
 
         Use this method instead of using the constructor to create a new sandbox.
         """
+        if auto_pause is not None:
+            warnings.warn(
+                "`auto_pause` is deprecated; use `lifecycle={'on_timeout': 'pause'}` instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+
         if not template and mcp is not None:
             template = cls.default_mcp_template
         elif not template:
@@ -606,9 +616,7 @@ class AsyncSandbox(SandboxApi):
             secure=secure,
             allow_internet_access=allow_internet_access,
             mcp=mcp,
-            lifecycle=(
-                {"on_timeout": "pause", "auto_resume": False} if auto_pause else None
-            ),
+            lifecycle=lifecycle,
             volume_mounts=transformed_mounts,
             **opts,
         )
